refactor(onboarding): Extract NewWelcomeUI to separate component file
Move NewWelcomeUI and its related components to a dedicated file
in the components folder for better code organization.
